Update context URLs
The WoT TD spec uses "2022" as the year so for consistency we have used that as the namespace for Discovery also.

Yes, "ns" would make more sense but we at least want Discovery and TD context files to be consistent, and the WoT TD files are already published.  Publication of WoT Discovery ontology files are pending.